+++ R/autoplot.R
@@ -8,6 +8,11 @@
 #' @param label_rotation Enables rotation of the treatment group labels independently of the x axis labels within the plot.
 #' @param margin Logical (default `FALSE`). A value of `FALSE` will expand the plot to the edges of the plotting area i.e. remove white space between plot and axes.
 #' @param palette A string specifying the colour scheme to use for plotting. Default is equivalent to "Spectral". Colour blind friendly palettes can also be provided via options `"colour blind"` (or `"color blind"`, both equivalent to `"viridis"`), `"magma"`, `"inferno"`, `"plasma"` or `"cividis"`. Other palettes from [scales::brewer_pal()] are also possible.
+#' @param buffer A string specifying the buffer plots to include for plotting. Default is `NULL` (no buffers plotted). Other options are "edge" (outer edge of trial area), "rows" (between rows), "columns" (between columns), "double row" (a buffer row each side of a treatment row) or "double column" (a buffer row each side of a treatment column). "blocks" (a buffer around each treatment block) will be implemented in a future release.
+#' @param row A variable to plot a column from `object` as rows.
+#' @param column A variable to plot a column from `object` as columns.
+#' @param block A variable to plot a column from `object` as blocks.
+#' @param treatments A variable to plot a column from `object` as treatments.
 #' @inheritParams rlang::args_dots_used
 #'
 #' @name autoplot
@@ -74,12 +79,13 @@ autoplot.mct <- function(object, size = 4, label_height = 0.1, rotation = 0, axi
 
 
 #' @rdname autoplot
+#'
 #' @importFrom farver decode_colour
 #' @importFrom grDevices colorRampPalette
 #' @importFrom ggplot2 ggplot geom_tile aes geom_text theme_bw scale_fill_manual scale_x_continuous scale_y_continuous scale_y_reverse
 #' @importFrom scales brewer_pal reverse_trans viridis_pal
 #' @importFrom stringi stri_sort
-#' @importFrom rlang check_dots_used
+#' @importFrom rlang check_dots_used enquo sym quo_is_null quo_name
 #' @export
 #' @examples
 #' des.out <- design(type = "crd", treatments = c(1, 5, 10, 20),
@@ -91,14 +97,40 @@ autoplot.mct <- function(object, size = 4, label_height = 0.1, rotation = 0, axi
 #'
 #' # Alternative colour scheme
 #' autoplot(des.out, palette = "plasma")
-autoplot.design <- function(object, rotation = 0, size = 4, margin = FALSE, palette = "default", row = NULL, col = NULL, block = NULL, ...) {
+autoplot.design <- function(object, rotation = 0, size = 4, margin = FALSE, palette = "default", buffer = NULL, row = NULL, column = NULL, block = NULL, treatments = NULL, ...) {
     stopifnot(inherits(object, "design"))
+    rlang::check_dots_used()
 
     if(inherits(object, "list")) {
         object <- object$design
     }
-    rlang::check_dots_used()
-    ntrt <- nlevels(as.factor(object$treatments))
+
+    row_expr <- rlang::enquo(row)
+    column_expr <- rlang::enquo(column)
+    block_expr <- rlang::enquo(block)
+    trt_expr <- rlang::enquo(treatments)
+
+    # If row and column are not provided, set default values
+    if(rlang::quo_is_null(row_expr)) {
+        row_expr <- rlang::sym("row")  # Default to the row column
+    }
+    if(rlang::quo_is_null(column_expr)) {
+        column_expr <- rlang::sym("col")  # Default to the col column
+    }
+    if(rlang::quo_is_null(block_expr)) {
+        block_expr <- rlang::sym("block")  # Default to the col column
+    }
+    if(rlang::quo_is_null(trt_expr)) {
+        trt_expr <- rlang::sym("treatments")  # Default to the col column
+    }
+
+    row_expr <- rlang::quo_name(row_expr)
+    column_expr <- rlang::quo_name(column_expr)
+    block_expr <- rlang::quo_name(block_expr)
+    trt_expr <- rlang::quo_name(trt_expr)
+
+    object[[trt_expr]] <- factor(as.character(object[[trt_expr]]), levels = unique(stringi::stri_sort(as.character(object[[trt_expr]]), numeric = TRUE)))
+    ntrt <- nlevels(object[[trt_expr]])
 
     # create the colours for the graph
     if(palette == "default") {
@@ -110,44 +142,48 @@ autoplot.design <- function(object, rotation = 0, size = 4, margin = FALSE, pale
     }
     else if(any(grepl("(colou?r([[:punct:]]|[[:space:]]?)blind)|cb|viridis", palette, ignore.case = T))) {
         colour_palette <- scales::viridis_pal(option = "viridis")(ntrt)
-        # Set text colour to be light on dark colours
-        # hcl <- farver::decode_colour(colour_palette, "rgb", "hcl")
-        # cols <- data.frame(treatments = levels(as.factor(object$treatments)),
-        #                    text_col = ifelse(hcl[, "l"] > 50, "black", "white"))
-        # object <- merge(object, cols)
     }
     else if(tolower(trimws(palette)) %in% c("magma", "inferno", "cividis", "plasma", "rocket", "mako", "turbo")) {
         colour_palette <- scales::viridis_pal(option = palette)(ntrt)
-        # Set text colour to be light on dark colours
-        # hcl <- farver::decode_colour(colour_palette, "rgb", "hcl")
-        # cols <- data.frame(treatments = levels(as.factor(object$treatments)),
-        #                    text_col = ifelse(hcl[, "l"] > 50, "black", "white"))
-        # object <- merge(object, cols)
     }
     else {
         stop("Invalid value for palette.")
     }
+
     hcl <- farver::decode_colour(colour_palette, "rgb", "hcl")
-    cols <- data.frame(treatments = levels(as.factor(object$treatments)),
+    colours <- data.frame(treatments = levels(object[[trt_expr]]),
                        text_col = ifelse(hcl[, "l"] > 50, "black", "white"))
-    object <- merge(object, cols)
+    colnames(colours)[1] <- trt_expr
+    object <- merge(object, colours)
 
     if(!any(grepl("block", tolower(names(object))))) {
+        if(!missing(buffer)) {
+            object <- create_buffers(object, type = buffer)
+            if("buffer" %in% levels(object[[trt_expr]])) {
+               colour_palette <- c(colour_palette, "white")
+            }
+        }
+
         # create the graph
         plt <- ggplot2::ggplot() +
-            ggplot2::geom_tile(data = object, mapping = ggplot2::aes(x = col, y = row, fill = treatments), colour = "black") +
-            ggplot2::geom_text(data = object, mapping = ggplot2::aes(x = col, y = row, label = treatments), colour = object$text_col, angle = rotation, size = size, ...) +
+            ggplot2::geom_tile(data = object, mapping = ggplot2::aes(x = .data[[column_expr]], y = .data[[row_expr]], fill = .data[[trt_expr]]), colour = "black") +
+            ggplot2::geom_text(data = object, mapping = ggplot2::aes(x = .data[[column_expr]], y = .data[[row_expr]], label = .data[[trt_expr]]), colour = object$text_col, angle = rotation, size = size, ...) +
             ggplot2::theme_bw()
     }
     else {
         # Set up dataframe with coordinates for drawing the blocks
         blkdf <- data.frame(
-            block = sort(unique(object$block)),
+            block = sort(unique(object[[block_expr]])),
             xmin = 0, xmax = 0, ymin = 0, ymax = 0
         )
+        if(!missing(buffer)) {
+            object <- create_buffers(object, type = buffer, blocks = TRUE)
+            if("buffer" %in% levels(object[[trt_expr]])) {
+                colour_palette <- c(colour_palette, "white")
+            }
+        }
         for (i in 1:nrow(blkdf)) {
-            item <- blkdf$block[i]
-            tmp <- object[object$block == item, ]
+            tmp <- object[object[[block_expr]] == blkdf$block[i], ]
             blkdf[i, "ymin"] <- (min(tmp$row) - 0.5)
             blkdf[i, "ymax"] <- (max(tmp$row) + 0.5)
             blkdf[i, "xmin"] <- (min(tmp$col) - 0.5)
@@ -155,8 +191,8 @@ autoplot.design <- function(object, rotation = 0, size = 4, margin = FALSE, pale
         }
 
         plt <- ggplot2::ggplot(...) +
-            ggplot2::geom_tile(data = object, mapping = ggplot2::aes(x = col, y = row, fill = treatments), colour = "black", ...) +
-            ggplot2::geom_text(data = object, mapping = ggplot2::aes(x = col, y = row, label = treatments), colour = object$text_col, angle = rotation, size = size, ...) +
+            ggplot2::geom_tile(data = object, mapping = ggplot2::aes(x = .data[[column_expr]], y = .data[[row_expr]], fill = .data[[trt_expr]]), colour = "black") +
+            ggplot2::geom_text(data = object, mapping = ggplot2::aes(x = .data[[column_expr]], y = .data[[row_expr]], label = .data[[trt_expr]]), colour = object$text_col, angle = rotation, size = size, ...) +
             ggplot2::geom_rect(
                 data = blkdf,
                 mapping = ggplot2::aes(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax),
@@ -173,10 +209,10 @@ autoplot.design <- function(object, rotation = 0, size = 4, margin = FALSE, pale
     plt <- plt + scale_fill_manual(values = colour_palette, name = "Treatment")
 
     if(!margin) {
-        plt <- plt + ggplot2::scale_x_continuous(expand = c(0, 0), breaks = seq(1, max(object$col), 1)) + ggplot2::scale_y_continuous(expand = c(0, 0), trans = scales::reverse_trans(), breaks = seq(1, max(object$row), 1))
+        plt <- plt + ggplot2::scale_x_continuous(expand = c(0, 0), breaks = seq(1, max(object[[column_expr]]), 1)) + ggplot2::scale_y_continuous(expand = c(0, 0), trans = scales::reverse_trans(), breaks = seq(1, max(object[[row_expr]]), 1))
     }
     else {
-        plt <- plt + ggplot2::scale_x_continuous(breaks = seq(1, max(object$col), 1))+ ggplot2::scale_y_continuous(trans = scales::reverse_trans(), breaks = seq(1, max(object$row), 1))
+        plt <- plt + ggplot2::scale_x_continuous(breaks = seq(1, max(object[[column_expr]]), 1))+ ggplot2::scale_y_continuous(trans = scales::reverse_trans(), breaks = seq(1, max(object[[row_expr]]), 1))
     }
 
     return(plt)