+++ R/logltest.R
@@ -88,18 +88,12 @@ logl_test <- function(model.obj, rand.terms = NULL, resid.terms = NULL, decimals
             while (!model.obj$converge & (n[1] < 10)) {
                 model.obj <- quiet(update(model.obj))
                 n[1] <- n[1] + 1
-                # if(n[1]==10) {
-                #     # warns <- c(warns, 'Model did not converge')
-                # }
             }
 
             n[2] <- 1
             while (any(model.obj$vparameters.pc > 1) & n[2] < 10) {
                 model.obj <- quiet(update(model.obj, ))
                 n[2] <- n[2] + 1
-                # if(n[2]==10) {
-                #     # warns <- c(warns, 'Model did not converge')
-                # }
             }
         }
 
@@ -118,18 +112,12 @@ logl_test <- function(model.obj, rand.terms = NULL, resid.terms = NULL, decimals
                 while (!model.obj1$converge & n[3] < 10) {
                     model.obj1 <- quiet(update(model.obj1))
                     n[3] <- n[3] + 1
-                    # if(n[3]==10) {
-                    #     warns <- c(warns, 'Model did not converge')
-                    # }
                 }
 
                 n[4] <- 1
                 while (!model.obj1$converge & n[4] < 10) {
                     model.obj1 <- quiet(update(model.obj1))
                     n[4] <- n[4] + 1
-                    # if(n[4]==10) {
-                    #     warns <- c(warns, 'Model did not converge')
-                    # }
                 }
 
                 # Logl test
@@ -156,18 +144,12 @@ logl_test <- function(model.obj, rand.terms = NULL, resid.terms = NULL, decimals
                 while (!model.obj1$converge & n[5] < 10) {
                     model.obj1 <- quiet(update(model.obj1))
                     n[5] <- n[5] + 1
-                    # if(n[5]==10) {
-                    #     warns <- c(warns, 'Model did not converge')
-                    # }
                 }
 
                 n[6] <- 1
                 while (any(model.obj1$vparameters.pc > 1) & n[6] < 10) {
                     model.obj1 <- quiet(update(model.obj1))
                     n[6] <- n[6] + 1
-                    # if(n[6]==10) {
-                    #     warns <- c(warns, 'Model did not converge')
-                    # }
                 }
 
                 # Logl test
@@ -197,10 +179,6 @@ logl_test <- function(model.obj, rand.terms = NULL, resid.terms = NULL, decimals
 
     all.terms <- c(rand.terms, resid.terms)
     test.df <- test.df[is.element(test.df$Term, all.terms),]
-
-    # if(length(warns) > 0 & !quiet) {
-    #     warning(warns[1], call. = FALSE)
-    # }
 
     return(test.df)
 }